+++ core/src/test/scala/unit/kafka/server/LogRecoveryTest.scala
@@ -16,26 +16,24 @@
 */
 package kafka.server
 
-import java.util.Properties
-
-import scala.collection.Seq
-
-import kafka.utils.TestUtils
-import TestUtils._
-import java.io.File
-
 import kafka.server.checkpoints.OffsetCheckpointFile
+import kafka.utils.TestUtils
+import kafka.utils.TestUtils._
 import org.apache.kafka.clients.admin.Admin
 import org.apache.kafka.clients.producer.{KafkaProducer, ProducerRecord}
-import org.apache.kafka.common.network.ListenerName
 import org.apache.kafka.common.TopicPartition
+import org.apache.kafka.common.network.ListenerName
 import org.apache.kafka.common.security.auth.SecurityProtocol
 import org.apache.kafka.common.serialization.{IntegerSerializer, StringSerializer}
 import org.apache.kafka.server.config.ReplicationConfigs
-import org.junit.jupiter.api.{AfterEach, BeforeEach, TestInfo}
 import org.junit.jupiter.api.Assertions._
+import org.junit.jupiter.api.{AfterEach, BeforeEach, TestInfo}
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+
+import java.io.File
+import java.util.Properties
+import scala.collection.Seq
 
 class LogRecoveryTest extends QuorumTestHarness {
 
@@ -253,5 +251,21 @@ class LogRecoveryTest extends QuorumTestHarness {
 
   private def sendMessages(n: Int): Unit = {
     (0 until n).map(_ => producer.send(new ProducerRecord(topic, 0, message))).foreach(_.get)
+  }
+
+  private def getLeaderIdForPartition[B <: KafkaBroker](
+                                                 brokers: Seq[B],
+                                                 tp: TopicPartition,
+                                                 timeout: Long = org.apache.kafka.test.TestUtils.DEFAULT_MAX_WAIT_MS): Int = {
+    def leaderExists: Option[Int] = {
+      brokers.find { broker =>
+        broker.replicaManager.onlinePartition(tp).exists(_.leaderLogIfLocal.isDefined)
+      }.map(_.config.brokerId)
+    }
+
+    waitUntilTrue(() => leaderExists.isDefined,
+      s"Did not find a leader for partition $tp after $timeout ms", waitTimeMs = timeout)
+
+    leaderExists.get
   }
 }
