+++ core/src/test/scala/unit/kafka/server/ServerGenerateClusterIdTest.scala
@@ -22,6 +22,7 @@ import scala.concurrent._
 import scala.concurrent.duration._
 import ExecutionContext.Implicits._
 import kafka.utils.TestUtils
+import org.apache.kafka.common.utils.Time
 import org.apache.kafka.metadata.properties.{MetaProperties, MetaPropertiesEnsemble, MetaPropertiesVersion, PropertiesUtils}
 import org.junit.jupiter.api.Assertions._
 import org.junit.jupiter.api.{AfterEach, BeforeEach, Test, TestInfo}
@@ -56,7 +57,7 @@ class ServerGenerateClusterIdTest extends QuorumTestHarness {
     // Make sure that the cluster id doesn't exist yet.
     assertFalse(zkClient.getClusterId.isDefined)
 
-    var server1 = TestUtils.createServer(config1, threadNamePrefix = Option(this.getClass.getName))
+    var server1 = createServer(config1, threadNamePrefix = Option(this.getClass.getName))
     servers = Seq(server1)
 
     // Validate the cluster id
@@ -70,7 +71,7 @@ class ServerGenerateClusterIdTest extends QuorumTestHarness {
     assertEquals(zkClient.getClusterId, Some(clusterIdOnFirstBoot))
 
     // Restart the server check to confirm that it uses the clusterId generated previously
-    server1 = TestUtils.createServer(config1, threadNamePrefix = Option(this.getClass.getName))
+    server1 = createServer(config1, threadNamePrefix = Option(this.getClass.getName))
     servers = Seq(server1)
 
     val clusterIdOnSecondBoot = server1.clusterId
@@ -87,13 +88,13 @@ class ServerGenerateClusterIdTest extends QuorumTestHarness {
 
   @Test
   def testAutoGenerateClusterIdForKafkaClusterSequential(): Unit = {
-    val server1 = TestUtils.createServer(config1, threadNamePrefix = Option(this.getClass.getName))
+    val server1 = createServer(config1, threadNamePrefix = Option(this.getClass.getName))
     val clusterIdFromServer1 = server1.clusterId
 
-    val server2 = TestUtils.createServer(config2, threadNamePrefix = Option(this.getClass.getName))
+    val server2 = createServer(config2, threadNamePrefix = Option(this.getClass.getName))
     val clusterIdFromServer2 = server2.clusterId
 
-    val server3 = TestUtils.createServer(config3, threadNamePrefix = Option(this.getClass.getName))
+    val server3 = createServer(config3, threadNamePrefix = Option(this.getClass.getName))
     val clusterIdFromServer3 = server3.clusterId
     servers = Seq(server1, server2, server3)
 
@@ -117,7 +118,7 @@ class ServerGenerateClusterIdTest extends QuorumTestHarness {
 
   @Test
   def testAutoGenerateClusterIdForKafkaClusterParallel(): Unit = {
-    val firstBoot = Future.traverse(Seq(config1, config2, config3))(config => Future(TestUtils.createServer(config, threadNamePrefix = Option(this.getClass.getName))))
+    val firstBoot = Future.traverse(Seq(config1, config2, config3))(config => Future(createServer(config, threadNamePrefix = Option(this.getClass.getName))))
     servers = Await.result(firstBoot, 100.second)
     val Seq(server1, server2, server3) = servers
 
@@ -145,7 +146,7 @@ class ServerGenerateClusterIdTest extends QuorumTestHarness {
   @Test
   def testConsistentClusterIdFromZookeeperAndFromMetaProps(): Unit = {
     // Check at the first boot
-    val server = TestUtils.createServer(config1, threadNamePrefix = Option(this.getClass.getName))
+    val server = createServer(config1, threadNamePrefix = Option(this.getClass.getName))
     val clusterId = server.clusterId
 
     assertTrue(verifyBrokerMetadata(server.config.logDirs, clusterId))
@@ -227,5 +228,9 @@ class ServerGenerateClusterIdTest extends QuorumTestHarness {
       }
     }
     true
+  }
+
+  def createServer(config: KafkaConfig, threadNamePrefix: Option[String]): KafkaServer = {
+    TestUtils.createServer(config, Time.SYSTEM, threadNamePrefix)
   }
 }
