+++ core/src/test/scala/unit/kafka/utils/TestUtils.scala
@@ -16,23 +16,9 @@
  */
 package kafka.utils
 
-import java.io._
-import java.net.InetAddress
-import java.nio._
-import java.nio.channels._
-import java.nio.charset.{Charset, StandardCharsets}
-import java.nio.file.{Files, StandardOpenOption}
-import java.security.cert.X509Certificate
-import java.time.Duration
-import java.util
-import java.util.concurrent.atomic.{AtomicBoolean, AtomicInteger}
-import java.util.concurrent.{Callable, CompletableFuture, ExecutionException, Executors, TimeUnit}
-import java.util.{Arrays, Collections, Optional, Properties}
-import com.yammer.metrics.core.{Gauge, Histogram, Meter}
-
-import javax.net.ssl.X509TrustManager
+import com.yammer.metrics.core.{Histogram, Meter}
 import kafka.api._
-import kafka.cluster.{AlterPartitionListener, Broker, EndPoint}
+import kafka.cluster.AlterPartitionListener
 import kafka.controller.{ControllerEventManager, LeaderIsrAndControllerEpoch}
 import kafka.log._
 import kafka.network.RequestChannel
@@ -42,17 +28,16 @@ import kafka.server.metadata.{ConfigRepository, MockConfigRepository}
 import kafka.tools.StorageTool
 import kafka.utils.Implicits._
 import kafka.zk._
-import org.apache.kafka.admin.BrokerMetadata
-import org.apache.kafka.clients.{ClientResponse, CommonClientConfigs}
 import org.apache.kafka.clients.admin.AlterConfigOp.OpType
 import org.apache.kafka.clients.admin._
 import org.apache.kafka.clients.consumer._
 import org.apache.kafka.clients.consumer.internals.AbstractCoordinator
 import org.apache.kafka.clients.producer.{KafkaProducer, ProducerConfig, ProducerRecord}
-import org.apache.kafka.common.{KafkaFuture, Node, TopicIdPartition, TopicPartition, Uuid}
-import org.apache.kafka.common.acl.{AccessControlEntry, AccessControlEntryFilter, AclBinding, AclBindingFilter}
-import org.apache.kafka.common.config.{ConfigException, ConfigResource}
+import org.apache.kafka.clients.{ClientResponse, CommonClientConfigs}
+import org.apache.kafka.common._
+import org.apache.kafka.common.acl.{AccessControlEntry, AccessControlEntryFilter, AclBindingFilter}
 import org.apache.kafka.common.config.ConfigResource.Type.TOPIC
+import org.apache.kafka.common.config.{ConfigException, ConfigResource}
 import org.apache.kafka.common.errors.{KafkaStorageException, OperationNotAttemptedException, TopicExistsException, UnknownTopicOrPartitionException}
 import org.apache.kafka.common.header.Header
 import org.apache.kafka.common.internals.Topic
@@ -61,17 +46,16 @@ import org.apache.kafka.common.message.UpdateMetadataRequestData.UpdateMetadataP
 import org.apache.kafka.common.metrics.Metrics
 import org.apache.kafka.common.network.{ClientInformation, ListenerName, Mode}
 import org.apache.kafka.common.protocol.{ApiKeys, Errors}
-import org.apache.kafka.common.quota.{ClientQuotaAlteration, ClientQuotaEntity}
 import org.apache.kafka.common.record._
-import org.apache.kafka.common.requests.{AbstractRequest, AbstractResponse, EnvelopeRequest, FetchRequest, RequestContext, RequestHeader}
+import org.apache.kafka.common.requests._
 import org.apache.kafka.common.resource.ResourcePattern
 import org.apache.kafka.common.security.auth.{KafkaPrincipal, KafkaPrincipalSerde, SecurityProtocol}
-import org.apache.kafka.common.serialization.{ByteArrayDeserializer, ByteArraySerializer, Deserializer, IntegerSerializer, Serializer}
+import org.apache.kafka.common.serialization._
 import org.apache.kafka.common.utils.Utils.formatAddress
 import org.apache.kafka.common.utils.{Time, Utils}
 import org.apache.kafka.controller.QuorumController
-import org.apache.kafka.coordinator.transaction.TransactionLogConfigs
 import org.apache.kafka.coordinator.group.GroupCoordinatorConfig
+import org.apache.kafka.coordinator.transaction.TransactionLogConfigs
 import org.apache.kafka.metadata.properties.MetaProperties
 import org.apache.kafka.network.SocketServerConfigs
 import org.apache.kafka.raft.QuorumConfig
@@ -79,9 +63,11 @@ import org.apache.kafka.server.{ClientMetricsManager, ControllerRequestCompletio
 import org.apache.kafka.server.authorizer.{AuthorizableRequestContext, Authorizer => JAuthorizer}
 import org.apache.kafka.server.config.{KRaftConfigs, ReplicationConfigs, ServerLogConfigs, QuotaConfigs, ZkConfigs}
 import org.apache.kafka.server.common.{ApiMessageAndVersion, MetadataVersion}
+import org.apache.kafka.server.config.{QuotaConfigs, ReplicationConfigs, ServerLogConfigs, ZkConfigs}
 import org.apache.kafka.server.metrics.KafkaYammerMetrics
 import org.apache.kafka.server.util.MockTime
 import org.apache.kafka.server.util.timer.SystemTimer
+import org.apache.kafka.server.{ClientMetricsManager, ControllerRequestCompletionHandler}
 import org.apache.kafka.storage.internals.log.{CleanerConfig, LogConfig, LogDirFailureChannel, ProducerStateManagerConfig}
 import org.apache.kafka.test.{TestSslUtils, TestUtils => JTestUtils}
 import org.apache.zookeeper.KeeperException.SessionExpiredException
@@ -91,6 +77,16 @@ import org.junit.jupiter.api.Assertions._
 import org.mockito.ArgumentMatchers.{any, anyBoolean}
 import org.mockito.Mockito
 
+import java.io._
+import java.net.InetAddress
+import java.nio._
+import java.nio.charset.{Charset, StandardCharsets}
+import java.nio.file.{Files, StandardOpenOption}
+import java.time.Duration
+import java.util
+import java.util.concurrent._
+import java.util.concurrent.atomic.{AtomicBoolean, AtomicInteger}
+import java.util.{Arrays, Collections, Optional, Properties}
 import scala.annotation.nowarn
 import scala.collection.mutable.{ArrayBuffer, ListBuffer}
 import scala.collection.{Map, Seq, immutable, mutable}
@@ -133,8 +129,6 @@ object TestUtils extends Logging {
    */
   def tempDir(): File = JTestUtils.tempDirectory()
 
-  def tempTopic(): String = "testTopic" + random.nextInt(1000000)
-
   /**
    * Create a temporary relative directory
    */
@@ -169,20 +163,12 @@ object TestUtils extends Logging {
    */
   def tempFile(contents: String): File = JTestUtils.tempFile(contents)
 
-  def tempPropertiesFile(properties: Properties): File = {
-    return tempPropertiesFile(properties.asScala)
-  }
+  def tempPropertiesFile(properties: Properties): File = tempPropertiesFile(properties.asScala)
 
   def tempPropertiesFile(properties: Map[String, String]): File = {
     val content = properties.map{case (k, v) => k + "=" + v}.mkString(System.lineSeparator())
-    return tempFile(content)
+    tempFile(content)
   }
-
-  /**
-   * Create a temporary file and return an open file channel for this file
-   */
-  def tempChannel(): FileChannel =
-    FileChannel.open(tempFile().toPath, StandardOpenOption.READ, StandardOpenOption.WRITE)
 
   /**
    * Create a kafka server instance with appropriate test settings
@@ -192,10 +178,6 @@ object TestUtils extends Logging {
    */
   def createServer(config: KafkaConfig, time: Time = Time.SYSTEM): KafkaServer = {
     createServer(config, time, None)
-  }
-
-  def createServer(config: KafkaConfig, threadNamePrefix: Option[String]): KafkaServer = {
-    createServer(config, Time.SYSTEM, threadNamePrefix)
   }
 
   def createServer(config: KafkaConfig, time: Time, threadNamePrefix: Option[String]): KafkaServer = {
@@ -215,11 +197,6 @@ object TestUtils extends Logging {
 
   def boundPort(broker: KafkaBroker, securityProtocol: SecurityProtocol = SecurityProtocol.PLAINTEXT): Int =
     broker.boundPort(ListenerName.forSecurityProtocol(securityProtocol))
-
-  def createBrokerAndEpoch(id: Int, host: String, port: Int, securityProtocol: SecurityProtocol = SecurityProtocol.PLAINTEXT,
-                           epoch: Long = 0): (Broker, Long) = {
-    (new Broker(id, host, port, ListenerName.forSecurityProtocol(securityProtocol), securityProtocol), epoch)
-  }
 
   /**
    * Create a test config for the provided parameters.
@@ -476,7 +453,7 @@ object TestUtils extends Logging {
       cause != null &&
         cause.isInstanceOf[TopicExistsException] &&
         // wait until all partitions metadata are propagated before verifying partitions number and replication factor
-        !waitForAllPartitionsMetadata(brokers, topic, effectiveNumPartitions).isEmpty &&
+        waitForAllPartitionsMetadata(brokers, topic, effectiveNumPartitions).nonEmpty &&
         topicHasSameNumPartitionsAndReplicationFactor(admin, topic, effectiveNumPartitions, replicationFactor)
     }
 
@@ -504,33 +481,6 @@ object TestUtils extends Logging {
       i -> allPartitionsMetadata.get(new TopicPartition(topic, i)).map(_.leader()).getOrElse(
         throw new IllegalStateException(s"Cannot get the partition leader for topic: $topic, partition: $i in server metadata cache"))
     }.toMap
-  }
-
-  def increasePartitions[B <: KafkaBroker](admin: Admin,
-                                              topic: String,
-                                              totalPartitionCount: Int,
-                                              brokersToValidate: Seq[B]
-                                            ): Unit = {
-
-    try {
-      val newPartitionSet: Map[String, NewPartitions] = Map.apply(topic -> NewPartitions.increaseTo(totalPartitionCount))
-      admin.createPartitions(newPartitionSet.asJava)
-    } catch {
-      case e: ExecutionException =>
-        throw e
-    }
-
-    if (brokersToValidate.size > 0) {
-      // wait until we've propagated all partitions metadata to all brokers
-      val allPartitionsMetadata = waitForAllPartitionsMetadata(brokersToValidate, topic, totalPartitionCount)
-
-      (0 until totalPartitionCount - 1).map { i =>
-        i -> allPartitionsMetadata.get(new TopicPartition(topic, i)).map(_.leader()).getOrElse(
-          throw new IllegalStateException(s"Cannot get the partition leader for topic: $topic, partition: $i in server metadata cache"))
-      }.toMap
-    } else {
-      Map.empty
-    }
   }
 
   def describeTopic(
@@ -665,42 +615,6 @@ object TestUtils extends Logging {
   }
 
   /**
-   * Ensures that the consumer offsets/group metadata topic exists. If it does not, the topic is created and the method waits
-   * until the leader is elected and metadata is propagated to all brokers. If it does, the method verifies that it has
-   * the expected number of partition and replication factor however it does not guarantee that the topic is empty.
-   */
-  def createOffsetsTopic(zkClient: KafkaZkClient, servers: Seq[KafkaBroker]): Unit = {
-    val server = servers.head
-    val numPartitions = server.config.offsetsTopicPartitions
-    val replicationFactor = server.config.offsetsTopicReplicationFactor.toInt
-
-    try {
-      createTopic(
-        zkClient,
-        Topic.GROUP_METADATA_TOPIC_NAME,
-        numPartitions,
-        replicationFactor,
-        servers,
-        server.groupCoordinator.groupMetadataTopicConfigs
-      )
-    } catch {
-      case ex: TopicExistsException =>
-        val allPartitionsMetadata = waitForAllPartitionsMetadata(
-          servers,
-          Topic.GROUP_METADATA_TOPIC_NAME,
-          numPartitions
-        )
-
-        // If the topic already exists, we ensure that it has the required
-        // number of partitions and replication factor. If it has not, the
-        // exception is thrown further.
-        if (allPartitionsMetadata.size != numPartitions || allPartitionsMetadata.head._2.replicas.size != replicationFactor) {
-          throw ex
-        }
-    }
-  }
-
-  /**
    * Wrap a single record log buffer.
    */
   def singletonRecords(value: Array[Byte],
@@ -709,12 +623,6 @@ object TestUtils extends Logging {
                        timestamp: Long = RecordBatch.NO_TIMESTAMP,
                        magicValue: Byte = RecordBatch.CURRENT_MAGIC_VALUE): MemoryRecords = {
     records(Seq(new SimpleRecord(timestamp, key, value)), magicValue = magicValue, codec = codec)
-  }
-
-  def recordsWithValues(magicValue: Byte,
-                        codec: CompressionType,
-                        values: Array[Byte]*): MemoryRecords = {
-    records(values.map(value => new SimpleRecord(value)), magicValue, codec)
   }
 
   def records(records: Iterable[SimpleRecord],
@@ -738,87 +646,6 @@ object TestUtils extends Logging {
    * @param numBytes The size of the array
    */
   def randomBytes(numBytes: Int): Array[Byte] = JTestUtils.randomBytes(numBytes)
-
-  /**
-   * Generate a random string of letters and digits of the given length
-   *
-   * @param len The length of the string
-   * @return The random string
-   */
-  def randomString(len: Int): String = JTestUtils.randomString(len)
-
-  /**
-   * Check that the buffer content from buffer.position() to buffer.limit() is equal
-   */
-  def checkEquals(b1: ByteBuffer, b2: ByteBuffer): Unit = {
-    assertEquals(b1.limit() - b1.position(), b2.limit() - b2.position(), "Buffers should have equal length")
-    for (i <- 0 until b1.limit() - b1.position())
-      assertEquals(b1.get(b1.position() + i), b2.get(b1.position() + i), "byte " + i + " byte not equal.")
-  }
-
-  /**
-   *  Throw an exception if an iterable has different length than expected
-   *
-   */
-  def checkLength[T](s1: Iterator[T], expectedLength:Int): Unit = {
-    var n = 0
-    while (s1.hasNext) {
-      n += 1
-      s1.next()
-    }
-    assertEquals(expectedLength, n)
-  }
-
-  /**
-   * Throw an exception if the two iterators are of differing lengths or contain
-   * different messages on their Nth element
-   */
-  def checkEquals[T](s1: java.util.Iterator[T], s2: java.util.Iterator[T]): Unit = {
-    while (s1.hasNext && s2.hasNext)
-      assertEquals(s1.next, s2.next)
-    assertFalse(s1.hasNext, "Iterators have uneven length--first has more")
-    assertFalse(s2.hasNext, "Iterators have uneven length--second has more")
-  }
-
-  def stackedIterator[T](s: Iterator[T]*): Iterator[T] = {
-    new Iterator[T] {
-      var cur: Iterator[T] = _
-      val topIterator = s.iterator
-
-      def hasNext: Boolean = {
-        while (true) {
-          if (cur == null) {
-            if (topIterator.hasNext)
-              cur = topIterator.next()
-            else
-              return false
-          }
-          if (cur.hasNext)
-            return true
-          cur = null
-        }
-        // should never reach here
-        throw new RuntimeException("should not reach here")
-      }
-
-      def next() : T = cur.next()
-    }
-  }
-
-  /**
-   * Create a hexadecimal string for the given bytes
-   */
-  def hexString(bytes: Array[Byte]): String = hexString(ByteBuffer.wrap(bytes))
-
-  /**
-   * Create a hexadecimal string for the given bytes
-   */
-  def hexString(buffer: ByteBuffer): String = {
-    val builder = new StringBuilder("0x")
-    for (i <- 0 until buffer.limit())
-      builder.append(String.format("%x", Integer.valueOf(buffer.get(buffer.position() + i))))
-    builder.toString
-  }
 
   /**
    * Returns security configuration options for broker or clients
@@ -932,21 +759,6 @@ object TestUtils extends Logging {
     consumerProps.put(ConsumerConfig.ISOLATION_LEVEL_CONFIG, if (readCommitted) "read_committed" else "read_uncommitted")
     consumerProps ++= consumerSecurityConfigs(securityProtocol, trustStoreFile, saslProperties)
     new KafkaConsumer[K, V](consumerProps, keyDeserializer, valueDeserializer)
-  }
-
-  def createBrokersInZk(zkClient: KafkaZkClient, ids: Seq[Int]): Seq[Broker] =
-    createBrokersInZk(ids.map(new BrokerMetadata(_, Optional.empty())), zkClient)
-
-  def createBrokersInZk(brokerMetadatas: Seq[BrokerMetadata], zkClient: KafkaZkClient): Seq[Broker] = {
-    zkClient.makeSurePersistentPathExists(BrokerIdsZNode.path)
-    val brokers = brokerMetadatas.map { b =>
-      val protocol = SecurityProtocol.PLAINTEXT
-      val listenerName = ListenerName.forSecurityProtocol(protocol)
-      Broker(b.id, Seq(EndPoint("localhost", 6667, listenerName, protocol)), if (b.rack.isPresent) Some(b.rack.get()) else None)
-    }
-    brokers.foreach(b => zkClient.registerBroker(BrokerInfo(Broker(b.id, b.endPoints, rack = b.rack),
-      MetadataVersion.latestTesting, jmxPort = -1)))
-    brokers
   }
 
   def getMsgStrings(n: Int): Seq[String] = {
@@ -1124,24 +936,18 @@ object TestUtils extends Logging {
     }, msg = msg, pause = 0L, waitTimeMs = waitTimeMs)
   }
 
-  /**
-   * Wait for the presence of an optional value.
-   *
-   * @param func The function defining the optional value
-   * @param msg Error message in the case that the value never appears
-   * @param waitTimeMs Maximum time to wait
-   * @return The unwrapped value returned by the function
-   */
-  def awaitValue[T](func: () => Option[T], msg: => String, waitTimeMs: Long = JTestUtils.DEFAULT_MAX_WAIT_MS): T = {
-    var value: Option[T] = None
-    waitUntilTrue(() => {
-      value = func()
-      value.isDefined
-    }, msg, waitTimeMs)
-    value.get
+  def subscribeAndWaitForRecords(topic: String,
+                                 consumer: Consumer[Array[Byte], Array[Byte]],
+                                 waitTimeMs: Long = JTestUtils.DEFAULT_MAX_WAIT_MS): Unit = {
+    consumer.subscribe(Collections.singletonList(topic))
+    pollRecordsUntilTrue(
+      consumer,
+      (records: ConsumerRecords[Array[Byte], Array[Byte]]) => !records.isEmpty,
+      "Expected records",
+      waitTimeMs)
   }
 
-  /**
+   /**
     * Wait until the given condition is true or throw an exception if the given wait time elapses.
     *
     * @param condition condition to check
@@ -1164,7 +970,7 @@ object TestUtils extends Logging {
     throw new RuntimeException("unexpected error")
   }
 
-  /**
+   /**
     * Invoke `compute` until `predicate` is true or `waitTime` elapses.
     *
     * Return the last `compute` result and a boolean indicating whether `predicate` succeeded for that value.
@@ -1209,10 +1015,6 @@ object TestUtils extends Logging {
     }
   }
 
-  def isLeaderLocalOnBroker(topic: String, partitionId: Int, broker: KafkaBroker): Boolean = {
-    broker.replicaManager.onlinePartition(new TopicPartition(topic, partitionId)).exists(_.leaderLogIfLocal.isDefined)
-  }
-
   def findLeaderEpoch(brokerId: Int,
                       topicPartition: TopicPartition,
                       brokers: Iterable[KafkaBroker]): Int = {
@@ -1235,7 +1037,7 @@ object TestUtils extends Logging {
       .getOrElse(throw new AssertionError(s"Unable to locate follower for $topicPartition"))
   }
 
-  /**
+   /**
     * Wait until all brokers know about each other.
     *
     * @param brokers The Kafka brokers.
@@ -1330,11 +1132,6 @@ object TestUtils extends Logging {
     controllerId.getOrElse(throw new AssertionError(s"Controller not elected after $timeout ms"))
   }
 
-  def waitUntilQuorumLeaderElected(controllerServer: ControllerServer, timeout: Long = JTestUtils.DEFAULT_MAX_WAIT_MS): Int = {
-    val (leaderAndEpoch, _) = computeUntilTrue(controllerServer.raftManager.leaderAndEpoch, waitTime = timeout)(_.leaderId().isPresent)
-    leaderAndEpoch.leaderId().orElseThrow(() => new AssertionError(s"Quorum Controller leader not elected after $timeout ms"))
-  }
-
   def awaitLeaderChange[B <: KafkaBroker](
       brokers: Seq[B],
       tp: TopicPartition,
@@ -1353,22 +1150,6 @@ object TestUtils extends Logging {
     newLeaderExists.get
   }
 
-  def getLeaderIdForPartition[B <: KafkaBroker](
-      brokers: Seq[B],
-      tp: TopicPartition,
-      timeout: Long = JTestUtils.DEFAULT_MAX_WAIT_MS): Int = {
-    def leaderExists: Option[Int] = {
-      brokers.find { broker =>
-          broker.replicaManager.onlinePartition(tp).exists(_.leaderLogIfLocal.isDefined)
-      }.map(_.config.brokerId)
-    }
-
-    waitUntilTrue(() => leaderExists.isDefined,
-      s"Did not find a leader for partition $tp after $timeout ms", waitTimeMs = timeout)
-
-    leaderExists.get
-  }
-
   def waitUntilLeaderIsKnown[B <: KafkaBroker](
       brokers: Seq[B],
       tp: TopicPartition,
@@ -1385,49 +1166,12 @@ object TestUtils extends Logging {
     leaderIfExists.get
   }
 
-  def writeNonsenseToFile(fileName: File, position: Long, size: Int): Unit = {
-    val file = new RandomAccessFile(fileName, "rw")
-    file.seek(position)
-    for (_ <- 0 until size)
-      file.writeByte(random.nextInt(255))
-    file.close()
-  }
-
   def appendNonsenseToFile(file: File, size: Int): Unit = {
     val outputStream = Files.newOutputStream(file.toPath(), StandardOpenOption.APPEND)
     try {
       for (_ <- 0 until size)
         outputStream.write(random.nextInt(255))
     } finally outputStream.close()
-  }
-
-  def checkForPhantomInSyncReplicas(zkClient: KafkaZkClient, topic: String, partitionToBeReassigned: Int, assignedReplicas: Seq[Int]): Unit = {
-    val inSyncReplicas = zkClient.getInSyncReplicasForPartition(new TopicPartition(topic, partitionToBeReassigned))
-    // in sync replicas should not have any replica that is not in the new assigned replicas
-    val phantomInSyncReplicas = inSyncReplicas.get.toSet -- assignedReplicas.toSet
-    assertTrue(phantomInSyncReplicas.isEmpty,
-      "All in sync replicas %s must be in the assigned replica list %s".format(inSyncReplicas, assignedReplicas))
-  }
-
-  def ensureNoUnderReplicatedPartitions(zkClient: KafkaZkClient, topic: String, partitionToBeReassigned: Int, assignedReplicas: Seq[Int],
-                                                servers: Seq[KafkaServer]): Unit = {
-    val topicPartition = new TopicPartition(topic, partitionToBeReassigned)
-    waitUntilTrue(() => {
-        val inSyncReplicas = zkClient.getInSyncReplicasForPartition(topicPartition)
-        inSyncReplicas.get.size == assignedReplicas.size
-      },
-      "Reassigned partition [%s,%d] is under replicated".format(topic, partitionToBeReassigned))
-    var leader: Option[Int] = None
-    waitUntilTrue(() => {
-        leader = zkClient.getLeaderForPartition(topicPartition)
-        leader.isDefined
-      },
-      "Reassigned partition [%s,%d] is unavailable".format(topic, partitionToBeReassigned))
-    waitUntilTrue(() => {
-        val leaderBroker = servers.filter(s => s.config.brokerId == leader.get).head
-        leaderBroker.replicaManager.underReplicatedPartitionCount == 0
-      },
-      "Reassigned partition [%s,%d] is under-replicated as reported by the leader %d".format(topic, partitionToBeReassigned, leader.get))
   }
 
   // Note: Call this method in the test itself, rather than the @AfterEach method.
@@ -1438,29 +1182,6 @@ object TestUtils extends Logging {
     }
     val threadCount = nonDaemonThreads.size
     assertEquals(0, threadCount, s"Found unexpected $threadCount NonDaemon threads=${nonDaemonThreads.map(t => t.getName).mkString(", ")}")
-  }
-
-  // Some threads are closed, but the state didn't reflect in the JVM immediately, so add some wait time for it
-  def assertNoNonDaemonThreadsWithWaiting(threadNamePrefix: String, waitTimeMs: Long = 500L): Unit = {
-    var nonDemonThreads: mutable.Set[Thread] = mutable.Set.empty[Thread]
-    waitUntilTrue(() => {
-      nonDemonThreads = Thread.getAllStackTraces.keySet.asScala.filter { t =>
-        !t.isDaemon && t.isAlive && t.getName.startsWith(threadNamePrefix)
-      }
-      0 == nonDemonThreads.size
-    }, s"Found unexpected ${nonDemonThreads.size} NonDaemon threads=${nonDemonThreads.map(t => t.getName).mkString(", ")}", waitTimeMs)
-  }
-
-  def numThreadsRunning(threadNamePrefix: String, isDaemon: Boolean): mutable.Set[Thread] = {
-    Thread.getAllStackTraces.keySet.asScala.filter { t =>
-      isDaemon && t.isAlive && t.getName.startsWith(threadNamePrefix)
-    }
-  }
-
-  def allThreadStackTraces(): String = {
-    Thread.getAllStackTraces.asScala.map { case (thread, stackTrace) =>
-      thread.getName + "\n\t" + stackTrace.toList.map(_.toString).mkString("\n\t")
-    }.mkString("\n")
   }
 
   def formatDirectories(
@@ -1595,10 +1316,6 @@ object TestUtils extends Logging {
     }
   }
 
-  def createIsrChangeListener(): MockAlterPartitionListener = {
-    new MockAlterPartitionListener()
-  }
-
   def produceMessages[B <: KafkaBroker](
       brokers: Seq[B],
       records: Seq[ProducerRecord[Array[Byte], Array[Byte]]],
@@ -1691,7 +1408,6 @@ object TestUtils extends Logging {
     ), "Failed to hard-delete the delete directory")
   }
 
-
   def causeLogDirFailure(failureType: LogDirFailureType, leaderBroker: KafkaBroker, partition: TopicPartition): File = {
     // Make log directory of the partition on the leader broker inaccessible by replacing it with a file
     val localLog = leaderBroker.replicaManager.localLogOrException(partition)
@@ -1724,12 +1440,6 @@ object TestUtils extends Logging {
     new String(bytes, encoding)
   }
 
-  def copyOf(props: Properties): Properties = {
-    val copy = new Properties()
-    copy ++= props
-    copy
-  }
-
   def sslConfigs(mode: Mode, clientCert: Boolean, trustStoreFile: Option[File], certAlias: String,
                  certCn: String = SslCertificateCn,
                  tlsProtocol: String = TestSslUtils.DEFAULT_TLS_PROTOCOL_FOR_TESTS): Properties = {
@@ -1748,20 +1458,6 @@ object TestUtils extends Logging {
     val sslProps = new Properties()
     sslConfigs.forEach { (k, v) => sslProps.put(k, v) }
     sslProps
-  }
-
-  // a X509TrustManager to trust self-signed certs for unit tests.
-  def trustAllCerts: X509TrustManager = {
-    val trustManager = new X509TrustManager() {
-      override def getAcceptedIssuers: Array[X509Certificate] = {
-        null
-      }
-      override def checkClientTrusted(certs: Array[X509Certificate], authType: String): Unit = {
-      }
-      override def checkServerTrusted(certs: Array[X509Certificate], authType: String): Unit = {
-      }
-    }
-    trustManager
   }
 
   def waitAndVerifyAcls(expected: Set[AccessControlEntry],
@@ -1801,7 +1497,7 @@ object TestUtils extends Logging {
     }
   }
 
-  private def secureZkPaths(zkClient: KafkaZkClient): Seq[String] = {
+  def secureZkPaths(zkClient: KafkaZkClient): Seq[String] = {
     def subPaths(path: String): Seq[String] = {
       if (zkClient.pathExists(path))
         path +: zkClient.getChildren(path).map(c => path + "/" + c).flatMap(subPaths)
@@ -1812,38 +1508,7 @@ object TestUtils extends Logging {
     topLevelPaths.flatMap(subPaths)
   }
 
-  /**
-   * Verifies that all secure paths in ZK are created with the expected ACL.
-   */
-  def verifySecureZkAcls(zkClient: KafkaZkClient, usersWithAccess: Int): Unit = {
-    secureZkPaths(zkClient).foreach(path => {
-      if (zkClient.pathExists(path)) {
-        val sensitive = ZkData.sensitivePath(path)
-        // usersWithAccess have ALL access to path. For paths that are
-        // not sensitive, world has READ access.
-        val aclCount = if (sensitive) usersWithAccess else usersWithAccess + 1
-        val acls = zkClient.getAcl(path)
-        assertEquals(aclCount, acls.size, s"Invalid ACLs for $path $acls")
-        acls.foreach(acl => isAclSecure(acl, sensitive))
-      }
-    })
-  }
-
-  /**
-   * Verifies that secure paths in ZK have no access control. This is
-   * the case when zookeeper.set.acl=false and no ACLs have been configured.
-   */
-  def verifyUnsecureZkAcls(zkClient: KafkaZkClient): Unit = {
-    secureZkPaths(zkClient).foreach(path => {
-      if (zkClient.pathExists(path)) {
-        val acls = zkClient.getAcl(path)
-        assertEquals(1, acls.size, s"Invalid ACLs for $path $acls")
-        acls.foreach(isAclUnsecure)
-      }
-    })
-  }
-
-  /**
+   /**
     * To use this you pass in a sequence of functions that are your arrange/act/assert test on the SUT.
     * They all run at the same time in the assertConcurrent method; the chances of triggering a multithreading code error,
     * and thereby failing some assertion are greatly increased.
@@ -1919,24 +1584,6 @@ object TestUtils extends Logging {
                            waitTimeMs: Long = JTestUtils.DEFAULT_MAX_WAIT_MS): Seq[ConsumerRecord[K, V]] = {
     val records = pollUntilAtLeastNumRecords(consumer, numRecords, waitTimeMs)
     assertEquals(numRecords, records.size, "Consumed more records than expected")
-    records
-  }
-
-  /**
-    * Will consume all the records for the given consumer for the specified duration. If you want to drain all the
-    * remaining messages in the partitions the consumer is subscribed to, the duration should be set high enough so
-    * that the consumer has enough time to poll everything. This would be based on the number of expected messages left
-    * in the topic, and should not be too large (ie. more than a second) in our tests.
-    *
-    * @return All the records consumed by the consumer within the specified duration.
-    */
-  def consumeRecordsFor[K, V](consumer: Consumer[K, V], duration: Long = JTestUtils.DEFAULT_MAX_WAIT_MS): Seq[ConsumerRecord[K, V]] = {
-    val startTime = System.currentTimeMillis()
-    val records = new ArrayBuffer[ConsumerRecord[K, V]]()
-    waitUntilTrue(() => {
-      records ++= consumer.poll(Duration.ofMillis(50)).asScala
-      System.currentTimeMillis() - startTime > duration
-    }, s"The timeout $duration was greater than the maximum wait time.")
     records
   }
 
@@ -2056,16 +1703,6 @@ object TestUtils extends Logging {
     adminClient.incrementalAlterConfigs(configs)
   }
 
-  def alterClientQuotas(adminClient: Admin, request: Map[ClientQuotaEntity, Map[String, Option[Double]]]): AlterClientQuotasResult = {
-    val entries = request.map { case (entity, alter) =>
-      val ops = alter.map { case (key, value) =>
-        new ClientQuotaAlteration.Op(key, value.map(Double.box).orNull)
-      }.asJavaCollection
-      new ClientQuotaAlteration(entity, ops)
-    }.asJavaCollection
-    adminClient.alterClientQuotas(entries)
-  }
-
   def assertLeader(client: Admin, topicPartition: TopicPartition, expectedLeader: Int): Unit = {
     waitForLeaderToBecome(client, topicPartition, Some(expectedLeader))
   }
@@ -2146,56 +1783,6 @@ object TestUtils extends Logging {
     )
   }
 
-  def waitForReplicasAssigned(client: Admin, partition: TopicPartition, brokerIds: Seq[Int]): Unit = {
-    waitUntilTrue(
-      () => {
-        val description = client.describeTopics(Set(partition.topic).asJava).allTopicNames.get.asScala
-        val replicas = description
-          .values
-          .flatMap(_.partitions.asScala.flatMap(_.replicas.asScala))
-          .map(_.id)
-          .toSeq
-
-        brokerIds == replicas
-      },
-      s"Expected brokers $brokerIds to be the replicas for $partition"
-    )
-  }
-
-  /**
-   * Capture the console output during the execution of the provided function.
-   */
-  def grabConsoleOutput(f: => Unit) : String = {
-    val out = new ByteArrayOutputStream
-    try scala.Console.withOut(out)(f)
-    finally scala.Console.out.flush()
-    out.toString
-  }
-
-  /**
-   * Capture the console error during the execution of the provided function.
-   */
-  def grabConsoleError(f: => Unit) : String = {
-    val err = new ByteArrayOutputStream
-    try scala.Console.withErr(err)(f)
-    finally scala.Console.err.flush()
-    err.toString
-  }
-
-  /**
-   * Capture both the console output and console error during the execution of the provided function.
-   */
-  def grabConsoleOutputAndError(f: => Unit) : (String, String) = {
-    val out = new ByteArrayOutputStream
-    val err = new ByteArrayOutputStream
-    try scala.Console.withOut(out)(scala.Console.withErr(err)(f))
-    finally {
-      scala.Console.out.flush()
-      scala.Console.err.flush()
-    }
-    (out.toString, err.toString)
-  }
-
   def assertFutureExceptionTypeEquals(future: KafkaFuture[_], clazz: Class[_ <: Throwable],
                                       expectedErrorMessage: Option[String] = None): Unit = {
     val cause = assertThrows(classOf[ExecutionException], () => future.get()).getCause
@@ -2229,15 +1816,6 @@ object TestUtils extends Logging {
     total.toLong
   }
 
-  def yammerGaugeValue[T](metricName: String): Option[T] = {
-    KafkaYammerMetrics.defaultRegistry.allMetrics.asScala
-      .filter { case (k, _) => k.getMBeanName.endsWith(metricName) }
-      .values
-      .headOption
-      .map(_.asInstanceOf[Gauge[T]])
-      .map(_.value)
-  }
-
   def meterCount(metricName: String): Long = {
     meterCountOpt(metricName).getOrElse(fail(s"Unable to find metric $metricName"))
   }
@@ -2263,14 +1841,6 @@ object TestUtils extends Logging {
   def clearYammerMetrics(): Unit = {
     for (metricName <- KafkaYammerMetrics.defaultRegistry.allMetrics.keySet.asScala)
       KafkaYammerMetrics.defaultRegistry.removeMetric(metricName)
-  }
-
-  def clearYammerMetricsExcept(names: Set[String]): Unit = {
-    for (metricName <- KafkaYammerMetrics.defaultRegistry.allMetrics.keySet.asScala) {
-      if (!names.contains(metricName.getMBeanName)) {
-        KafkaYammerMetrics.defaultRegistry.removeMetric(metricName)
-      }
-    }
   }
 
   def stringifyTopicPartitions(partitions: Set[TopicPartition]): String = {
@@ -2309,7 +1879,7 @@ object TestUtils extends Logging {
     resetBrokersThrottle(admin, brokerIds)
   }
 
-  /**
+   /**
     * Throttles all replication across the cluster.
     * @param adminClient is the adminClient to use for making connection with the cluster
     * @param brokerIds all broker ids in the cluster
@@ -2391,48 +1961,6 @@ object TestUtils extends Logging {
     override def requestVersion(): Int = 0
     override def clientId(): String = ""
     override def correlationId(): Int = 0
-  }
-
-  def addAndVerifyAcls[B <: KafkaBroker](
-    brokers: Seq[B],
-    acls: Set[AccessControlEntry],
-    resource: ResourcePattern,
-    controllers: Seq[ControllerServer] = Seq(),
-  ): Unit = {
-    val authorizerForWrite = pickAuthorizerForWrite(brokers, controllers)
-    val aclBindings = acls.map { acl => new AclBinding(resource, acl) }
-    authorizerForWrite.createAcls(anonymousAuthorizableContext, aclBindings.toList.asJava).asScala
-      .map(_.toCompletableFuture.get)
-      .foreach { result =>
-        result.exception.ifPresent { e => throw e }
-      }
-    val aclFilter = new AclBindingFilter(resource.toFilter, AccessControlEntryFilter.ANY)
-    (brokers.map(_.authorizer.get) ++ controllers.map(_.authorizer.get)).foreach {
-      authorizer => waitAndVerifyAcls(
-        authorizer.acls(aclFilter).asScala.map(_.entry).toSet ++ acls,
-        authorizer, resource)
-    }
-  }
-
-  def removeAndVerifyAcls[B <: KafkaBroker](
-    brokers: Seq[B],
-    acls: Set[AccessControlEntry],
-    resource: ResourcePattern,
-    controllers: Seq[ControllerServer] = Seq(),
-  ): Unit = {
-    val authorizerForWrite = pickAuthorizerForWrite(brokers, controllers)
-    val aclBindingFilters = acls.map { acl => new AclBindingFilter(resource.toFilter, acl.toFilter) }
-    authorizerForWrite.deleteAcls(anonymousAuthorizableContext, aclBindingFilters.toList.asJava).asScala
-      .map(_.toCompletableFuture.get)
-      .foreach { result =>
-        result.exception.ifPresent { e => throw e }
-      }
-    val aclFilter = new AclBindingFilter(resource.toFilter, AccessControlEntryFilter.ANY)
-    (brokers.map(_.authorizer.get) ++ controllers.map(_.authorizer.get)).foreach {
-      authorizer => waitAndVerifyAcls(
-        authorizer.acls(aclFilter).asScala.map(_.entry).toSet -- acls,
-        authorizer, resource)
-    }
   }
 
   def buildEnvelopeRequest(
