+++ core/src/test/scala/unit/kafka/zk/AdminZkClientTest.scala
@@ -16,20 +16,22 @@
  */
 package kafka.admin
 
-import java.util
-import java.util.{Optional, Properties}
+import kafka.cluster.{Broker, EndPoint}
 import kafka.controller.ReplicaAssignment
 import kafka.server.KafkaConfig._
 import kafka.server.{KafkaConfig, KafkaServer, QuorumTestHarness}
 import kafka.utils.CoreUtils._
 import kafka.utils.TestUtils._
 import kafka.utils.{Logging, TestUtils}
-import kafka.zk.{AdminZkClient, ConfigEntityTypeZNode, KafkaZkClient}
+import kafka.zk._
+import org.apache.kafka.admin.BrokerMetadata
 import org.apache.kafka.common.TopicPartition
 import org.apache.kafka.common.config.TopicConfig
 import org.apache.kafka.common.errors.{InvalidReplicaAssignmentException, InvalidTopicException, TopicExistsException}
 import org.apache.kafka.common.metrics.Quota
-import org.apache.kafka.server.common.AdminOperationException
+import org.apache.kafka.common.network.ListenerName
+import org.apache.kafka.common.security.auth.SecurityProtocol
+import org.apache.kafka.server.common.{AdminOperationException, MetadataVersion}
 import org.apache.kafka.server.config.{ConfigType, QuotaConfigs}
 import org.apache.kafka.storage.internals.log.LogConfig
 import org.apache.kafka.test.{TestUtils => JTestUtils}
@@ -37,8 +39,10 @@ import org.junit.jupiter.api.Assertions._
 import org.junit.jupiter.api.{AfterEach, Test}
 import org.mockito.Mockito.{mock, when}
 
-import scala.jdk.CollectionConverters._
+import java.util
+import java.util.{Optional, Properties}
 import scala.collection.{Map, Seq, immutable}
+import scala.jdk.CollectionConverters._
 
 class AdminZkClientTest extends QuorumTestHarness with Logging with RackAwareTest {
 
@@ -55,7 +59,7 @@ class AdminZkClientTest extends QuorumTestHarness with Logging with RackAwareTes
   @Test
   def testManualReplicaAssignment(): Unit = {
     val brokers = List(0, 1, 2, 3, 4)
-    TestUtils.createBrokersInZk(zkClient, brokers)
+    createBrokersInZk(zkClient, brokers)
 
     val topicConfig = new Properties()
 
@@ -114,7 +118,7 @@ class AdminZkClientTest extends QuorumTestHarness with Logging with RackAwareTes
     )
     val topic = "test"
     val topicConfig = new Properties()
-    TestUtils.createBrokersInZk(zkClient, List(0, 1, 2, 3, 4))
+    createBrokersInZk(zkClient, List(0, 1, 2, 3, 4))
     // create the topic
     adminZkClient.createTopicWithAssignment(topic, topicConfig, expectedReplicaAssignment)
     // create leaders for all partitions
@@ -132,7 +136,7 @@ class AdminZkClientTest extends QuorumTestHarness with Logging with RackAwareTes
   def testTopicCreationWithCollision(): Unit = {
     val topic = "test.topic"
     val collidingTopic = "test_topic"
-    TestUtils.createBrokersInZk(zkClient, List(0, 1, 2, 3, 4))
+    createBrokersInZk(zkClient, List(0, 1, 2, 3, 4))
     // create the topic
     adminZkClient.createTopic(topic, 3, 1)
 
@@ -167,7 +171,7 @@ class AdminZkClientTest extends QuorumTestHarness with Logging with RackAwareTes
   @Test
   def testConcurrentTopicCreation(): Unit = {
     val topic = "test-concurrent-topic-creation"
-    TestUtils.createBrokersInZk(zkClient, List(0, 1, 2, 3, 4))
+    createBrokersInZk(zkClient, List(0, 1, 2, 3, 4))
     val props = new Properties
     props.setProperty(TopicConfig.MIN_IN_SYNC_REPLICAS_CONFIG, "2")
     def createTopic(): Unit = {
@@ -335,7 +339,7 @@ class AdminZkClientTest extends QuorumTestHarness with Logging with RackAwareTes
     val brokerList = 0 to 5
     val rackInfo = Map(0 -> "rack1", 1 -> "rack2", 2 -> "rack2", 3 -> "rack1", 5 -> "rack3")
     val brokerMetadatas = toBrokerMetadata(rackInfo, brokersWithoutRack = brokerList.filterNot(rackInfo.keySet))
-    TestUtils.createBrokersInZk(brokerMetadatas.asScala.toSeq, zkClient)
+    createBrokersInZk(brokerMetadatas.asScala.toSeq, zkClient)
 
     val processedMetadatas1 = adminZkClient.getBrokerMetadatas(RackAwareMode.Disabled)
     assertEquals(brokerList, processedMetadatas1.map(_.id))
@@ -421,5 +425,20 @@ class AdminZkClientTest extends QuorumTestHarness with Logging with RackAwareTes
     adminZkClient.changeIpConfig("127.0.0.1", new Properties())
     val users = zkClient.getChildren(ConfigEntityTypeZNode.path(ConfigType.IP))
     assert(users.isEmpty)
+  }
+
+  private def createBrokersInZk(zkClient: KafkaZkClient, ids: Seq[Int]): Seq[Broker] =
+    createBrokersInZk(ids.map(new BrokerMetadata(_, Optional.empty())), zkClient)
+
+  private def createBrokersInZk(brokerMetadatas: Seq[BrokerMetadata], zkClient: KafkaZkClient): Seq[Broker] = {
+    zkClient.makeSurePersistentPathExists(BrokerIdsZNode.path)
+    val brokers = brokerMetadatas.map { b =>
+      val protocol = SecurityProtocol.PLAINTEXT
+      val listenerName = ListenerName.forSecurityProtocol(protocol)
+      Broker(b.id, Seq(EndPoint("localhost", 6667, listenerName, protocol)), if (b.rack.isPresent) Some(b.rack.get()) else None)
+    }
+    brokers.foreach(b => zkClient.registerBroker(BrokerInfo(Broker(b.id, b.endPoints, rack = b.rack),
+      MetadataVersion.latestTesting, jmxPort = -1)))
+    brokers
   }
 }
